Extend ltunicode script to cover XeTeX-specific concepts
The file now should cover everything provided by unicode-letters.tex from
the core XeTeX distribution. The code for XeTeX inter character classes 
and token registers is almost identical to the version supplied with XeTeX:
there is just a \hskip => \hspace change. The processed data file produced
by this sript is a bit longer than that in the core XeTeX files: the approach
here is to load the work into the script end rather than format building.
Other than that, things should be the same in outcome. 

Note that while this checkin only affects the results for XeTeX the file
ltunicode.ltx is usable with both XeTeX and LuaTeX, and in principle any
other Unicode engine which supports the extended allocators and Unicode
math extensions that XeTeX and LuaTeX both provide.

The data file the script produces is not yet added to the sources: this will
be the next stage of the process here.