Add mechanism for loading Unicode data
This adds the extracted data generated by ltunicode.dtx as "ltunicode.ltx"
plus the appropriate hooks to load this in ltfinal.dtx.

The Unicode data is loaded before the hyphenattion patterns are set up: this
is required for XeTeX and does no harm for LuaTeX. As the Unicode hyphenation
patterns shold be 'well-behaved' the entire 'reset the codes' part is skipped. 
Hopefully this is OK: life gets rather more complex if the reset has to be 
interleaved.

Versions left unchanged as there's been no release if any of this code and
we are likely to have quite a lot of check-ins to get everything right.