Enable extra primitives when LuaTeX is in use
At present, the scheme here is the same as used by plain LuaTeX: enable all
of the primitives with their 'natural' names and a subset with the prefix
\luatex... as that has been the approach taken to date by the format builders.

Very old LuaTeX versions use a different syntax for \directlua than the current
release. However, they also provide the primitives 'out of the box' so there
is no need for a version test.