Compress unicode-data.def
There is no change to the data in this checkin.

By storing only ranges for east Asian glyph data (\ID ...) and by writing
caseless letters only once (\l ... not \L ...) the file can be reduced to ~20%
of the former size. There is no obvious loss of performance. This also makes
it easier to compare to the old unicode-data.tex.