Add an initial model for testing TU encodings
As mentioned on the team list, I will aim to add this functionality
to l3build shortly so a separate .lua file will not be required.

Note that at present LuaTeX testing is not working here: as this is
a demo that's 'to be addressed'!

The test file here is a toy but shows things work.