Fix tests
There are currently two callbacks which do not exists in the LuaTeX
version on Travis. We still need to register them because they are
necessary for LuaHBTeX, so just silence the test for now. I do not want
to change the test to ignore these callbacks because this should no
longer be relevant once we reach TeXlive 2020 and the callbacks are
available on LuaTeX too.