Avoid local format files being found in this test
When running this test with l3build, at least two format files are
found:  the file build by l3build to test LaTeX and the one installed by
the TeX distribution, and this ends up with a full path in the log file
which is tricky to filter out.  A saner solution is to make kpsewhich
find no format file by overriding a format-specific string.