Remove path from internal file name representation (fixes #422)
A path can be written several different ways and still lead to the same
file in the filesystem, so the same package file can be loaded twice as
long as the path is typed differently (it doesn't even need to point to
a different file in the system).  The sane solution is to drop the file
path from all internal bookkeeping and only use it in filesystem lookup
operations.  This commit introduces `\@currpath` to store the file path
and `\@currname` now contains only the base name.  This also fixes #421
although #421 also requires some changes to `xkeyval` because it parses
`\@filelist` (which contains paths) and looks at `\opt@path/name.ext`.