time travel is hard
This commit (hopefully) fixes an issue that when rolling back to a set
of changes already present in the format, errors about "command already
defined" would be thrown all over.  The fix "simply" stops the loading
of a given `\IncludeInRelease` chunk if it is already in the format
(which would be useless anyway).