Correctly detect `\set@curr@file(@nosearch)` in graphics.sty
Instead of making `\set@curr@file@nosearch` a copy of `\set@curr@file`
when rolling back, make graphics.sty correctly detect which of both is
currently defined, and use the latest (either the `@nosearch` version if
it exists, or the regular one) as a copy in `\Gin@set@curr@file`.  This
makes the code a bit easier, and better supports Plain (and also is more
correct than having a `@nosearch` macro that _does_ search).