Prevent math grabbing inside \m@th
This is needed as there are places that \m@th is used *within* a math mode construct. For example, \@tabarray uses \m@th, but is itself
used to construct the array env. The result is that if one 're-activates' math grabbing inside \m@th, \mathchoice in particular 'shows up'.

We will need to examine how to deal with truly
nested math-inside-text-using-\m@th later.

A knock-on of this change is we no longer need
to do anything with \mathchoice, as that always
shows up only inside math mode or inside a \m@th construct.