Use .ini files for all format building
This moves LuaTeX from DVI to PDF mode.

There are a few .luatex.tlg files that likely would
disappear if we switch from e-TeX to pdfTeX as
the main engine setup.

For graphics, only e-TeX and XeTeX are now
tested: we would need to work quite hard to
switch to PDF mode here.