Regen data structure after rollback
The 2023/06/01 release does heavy changes on the underlying data
structures of hooks, which are incompatible with previous releases.
Thus, when rolling back one has to “undo” the data structure changes
without losing what's already stored.  The change has to happen only
in the specific case of rolling back from a date (equal to or) later
than 2023/06/01 to a date before that.