Use a one-level indirection for optimized commands
This means that they still have "\foo code" and no args at the top level:
that makes them easier to handle with patching, etc.

Args are only grabbed once and we are looking at very simple expansion, so there is no real need
to worry about performance.

The marker \@@_start_optimized: is retained:
we could just check for "\foo code" directly,
but this way there is a clear single-token
marker at the start of the def.