Tag new functions in lthooks.dtx (#1241)
* Tag new functions in lthooks.dtx

New functions were recently added. The online doc won't always reflect the actual distribution of the reader.
We tag the new functions as added to the 2023-06-01 release. This is not very pretty for L3 API because new and old functions are documented altogether. But this problem will disappear in a forthcoming release.

* `\lthooksdate` updated