Update l3kernel to 2025-04-14
Rollback of the allocation routine means we run into issues with a few
tests. That is going to be addressed in the 2025-06 release, but to allow us to do l3kernel releases right now we have to put up with a few problematic test outputs. As this is analysed and understandable,
that should not pose a longer-term problem (nor impact real users).